+++ Applications/Classification/jni/main_func.cpp
@@ -33,7 +33,7 @@
 #include <stdlib.h>
 #include <time.h>
 
-#include "layers.h"
+#include "layer.h"
 #include "neuralnet.h"
 #include "tensor.h"
