+++ jni/Android.mk
@@ -23,7 +23,10 @@ endif
 
 NNTRAINER_SRCS := $(NNTRAINER_ROOT)/nntrainer/src/neuralnet.cpp \
                   $(NNTRAINER_ROOT)/nntrainer/src/tensor.cpp \
-                  $(NNTRAINER_ROOT)/nntrainer/src/layers.cpp \
+                  $(NNTRAINER_ROOT)/nntrainer/src/layer.cpp \
+                  $(NNTRAINER_ROOT)/nntrainer/src/input_layer.cpp \
+                  $(NNTRAINER_ROOT)/nntrainer/src/fc_layer.cpp \
+                  $(NNTRAINER_ROOT)/nntrainer/src/bn_layer.cpp \
                   $(NNTRAINER_ROOT)/nntrainer/src/databuffer.cpp \
                   $(NNTRAINER_ROOT)/nntrainer/src/databuffer_func.cpp \
                   $(NNTRAINER_ROOT)/nntrainer/src/databuffer_file.cpp \
@@ -51,4 +54,3 @@ LOCAL_SRC_FILES     := $(NNTRAINER_SRCS) $(INIPARSER_SRCS)
 LOCAL_C_INCLUDES    += $(NNTRAINER_INCLUDES) $(INIPARSER_INCLUDES)
 
 include $(BUILD_SHARED_LIBRARY)
-
