+++ nntrainer/include/fc_layer.h
@@ -1,0 +1,160 @@
+/**
+ * Copyright (C) 2020 Samsung Electronics Co., Ltd. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * @file	fc_layer.h
+ * @date	14 May 2020
+ * @brief	This is Fully Connected Layer Class of Neural Network
+ * @see		https://github.com/nnstreamer/nntrainer
+ * @author	Jijoong Moon <jijoong.moon@samsung.com>
+ * @bug		No known bugs except for NYI items
+ *
+ */
+
+#ifndef __FC_LAYER_H__
+#define __FC_LAYER_H__
+#ifdef __cplusplus
+
+#include <fstream>
+#include <iostream>
+#include <optimizer.h>
+#include <tensor.h>
+#include <vector>
+#include <layer.h>
+
+namespace nntrainer {
+
+/**
+ * @class   FullyConnecedLayer
+ * @brief   fully connected layer
+ */
+class FullyConnectedLayer : public Layer {
+public:
+  /**
+   * @brief     Constructor of Fully Connected Layer
+   */
+  FullyConnectedLayer() : loss(0.0), cost(COST_UNKNOWN){};
+
+  /**
+   * @brief     Destructor of Fully Connected Layer
+   */
+  ~FullyConnectedLayer(){};
+
+  /**
+   * @brief     Read Weight & Bias Data from file
+   * @param[in] file input stream file
+   */
+  void read(std::ifstream &file);
+
+  /**
+   * @brief     Save Weight & Bias Data to file
+   * @param[in] file output stream file
+   */
+  void save(std::ofstream &file);
+
+  /**
+   * @brief     forward propagation with input
+   * @param[in] in Input Tensor from upper layer
+   * @retval    Activation(W x input + B)
+   */
+  Tensor forwarding(Tensor in, int &status);
+
+  /**
+   * @brief     foward propagation : return Input Tensor
+   *            It return Input as it is.
+   * @param[in] input input Tensor from lower layer.
+   * @param[in] output label Tensor.
+   * @retval    Activation(W x input + B)
+   */
+  Tensor forwarding(Tensor in, Tensor output, int &status);
+
+  /**
+   * @brief     back propagation
+   *            Calculate dJdB & dJdW & Update W & B
+   * @param[in] input Input Tensor from lower layer
+   * @param[in] iteration Number of Epoch for ADAM
+   * @retval    dJdB x W Tensor
+   */
+  Tensor backwarding(Tensor in, int iteration);
+
+  /**
+   * @brief     copy layer
+   * @param[in] l layer to copy
+   */
+  void copy(std::shared_ptr<Layer> l);
+
+  /**
+   * @brief     initialize layer
+   * @param[in] b batch size
+   * @param[in] h height
+   * @param[in] w width
+   * @param[in] last last layer
+   * @param[in] init_zero boolean to set Bias zero
+   * @param[in] wini Weight Initialization Scheme
+   * @retval #ML_ERROR_NONE Successful.
+   * @retval #ML_ERROR_INVALID_PARAMETER invalid parameter.
+   */
+  int initialize(int b, int h, int w, bool last, bool init_zero,
+                 WeightIniType wini);
+  /**
+   * @brief     get Loss value
+   */
+  float getLoss() { return loss; }
+
+  /**
+   * @brief     set cost function
+   * @param[in] c cost function type
+   * @retval #ML_ERROR_NONE Successful.
+   * @retval #ML_ERROR_INVALID_PARAMETER invalid parameter.
+   */
+  int setCost(CostType c);
+
+  /**
+   * @brief     set Property of layer
+   * @param[in] values values of property
+   * @retval #ML_ERROR_NONE Successful.
+   * @retval #ML_ERROR_INVALID_PARAMETER invalid parameter.
+   */
+  int setProperty(std::vector<std::string> values);
+
+  /**
+   * @brief     Property Enumeration
+   *            0. input shape : string
+   *            1. bias zero : bool
+   *            4. activation : bool
+   *            6. weight_decay : string (type)
+   *            7. weight_decay_lambda : float
+   */
+  enum class PropertyType {
+    bias_zero = 1,
+    activation = 4,
+    weight_decay = 6,
+    weight_decay_lambda = 7,
+    unit = 8,
+  };
+
+private:
+  /**
+   * @brief     update loss
+   * @param[in] l Tensor data to calculate
+   */
+  void updateLoss(Tensor l);
+
+  Tensor weight;
+  Tensor bias;
+  float loss;
+  CostType cost;
+};
+} // namespace nntrainer
+
+#endif /* __cplusplus */
+#endif /* __FC_LAYER_H__ */
