+++ nntrainer/include/input_layer.h
@@ -1,0 +1,159 @@
+/**
+ * Copyright (C) 2020 Samsung Electronics Co., Ltd. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * @file	input_layer.h
+ * @date	14 May 2020
+ * @brief	This is Input Layer Class of Neural Network
+ * @see		https://github.com/nnstreamer/nntrainer
+ * @author	Jijoong Moon <jijoong.moon@samsung.com>
+ * @bug		No known bugs except for NYI items
+ *
+ */
+
+#ifndef __INPUT_LAYER_H__
+#define __INPUT_LAYER_H__
+#ifdef __cplusplus
+
+#include <fstream>
+#include <iostream>
+#include <optimizer.h>
+#include <tensor.h>
+#include <vector>
+#include <layer.h>
+
+namespace nntrainer {
+
+/**
+ * @class   Input Layer
+ * @brief   Just Handle the Input of Network
+ */
+class InputLayer : public Layer {
+public:
+  /**
+   * @brief     Constructor of InputLayer
+   */
+  InputLayer() : normalization(false), standardization(false){};
+
+  /**
+   * @brief     Destructor of InputLayer
+   */
+  ~InputLayer(){};
+
+  /**
+   * @brief     No Weight data for this Input Layer
+   */
+  void read(std::ifstream &file){};
+
+  /**
+   * @brief     No Weight data for this Input Layer
+   */
+  void save(std::ofstream &file){};
+
+  /**
+   * @brief     It is back propagation of input layer.
+   *            It return Input as it is.
+   * @param[in] input input Tensor from lower layer.
+   * @param[in] iteration Epoch Number for ADAM
+   * @retval
+   */
+  Tensor backwarding(Tensor in, int iteration) { return input; };
+
+  /**
+   * @brief     foward propagation : return Input Tensor
+   *            It return Input as it is.
+   * @param[in] in input Tensor from lower layer.
+   * @retval    return Input Tensor
+   */
+  Tensor forwarding(Tensor in, int &status);
+
+  /**
+   * @brief     foward propagation : return Input Tensor
+   *            It return Input as it is.
+   * @param[in] in input Tensor from lower layer.
+   * @param[in] output label Tensor.
+   * @retval    return Input Tensor
+   */
+  Tensor forwarding(Tensor in, Tensor output, int &status) {
+    return forwarding(in, status);
+  };
+
+  /**
+   * @brief     Set Optimizer
+   * @param[in] opt optimizer
+   * @retval #ML_ERROR_NONE Successful.
+   * @retval #ML_ERROR_INVALID_PARAMETER invalid parameter.
+   */
+  int setOptimizer(Optimizer &opt);
+
+  /**
+   * @brief     Initializer of Input Layer
+   * @param[in] b batch size
+   * @param[in] h height
+   * @param[in] w width
+   * @param[in] last last layer
+   * @param[in] init_zero boolean to set Bias zero
+   * @param[in] wini Weight Initialization Scheme
+   * @retval #ML_ERROR_NONE Successful.
+   * @retval #ML_ERROR_INVALID_PARAMETER invalid parameter.
+   */
+  int initialize(int b, int h, int w, bool last, bool init_zero,
+                 WeightIniType wini);
+
+  /**
+   * @brief     Copy Layer
+   * @param[in] l layer to copy
+   */
+  void copy(std::shared_ptr<Layer> l);
+
+  /**
+   * @brief     set normalization
+   * @param[in] enable boolean
+   */
+  void setNormalization(bool enable) { this->normalization = enable; };
+
+  /**
+   * @brief     set standardization
+   * @param[in] enable boolean
+   */
+  void setStandardization(bool enable) { this->standardization = enable; };
+
+  /**
+   * @brief     set Property of layer
+   * @param[in] values values of property
+   * @retval #ML_ERROR_NONE Successful.
+   * @retval #ML_ERROR_INVALID_PARAMETER invalid parameter.
+   */
+  int setProperty(std::vector<std::string> values);
+
+  /**
+   * @brief     Property Enumeration
+   *            0. input shape : string
+   *            1. bias zero : bool
+   *            2. normalization : bool
+   *            3. normalization : bool
+   */
+  enum class PropertyType {
+    input_shape = 0,
+    bias_zero = 1,
+    normalization = 2,
+    standardization = 3
+  };
+
+private:
+  bool normalization;
+  bool standardization;
+};
+} // namespace nntrainer
+
+#endif /* __cplusplus */
+#endif /* __INPUT_LAYER_H__ */
